Extract integration testing tools into bravado.testing.integration_test module
