+++ setup.py
@@ -43,5 +43,10 @@ setup(
     extras_require={
         'fido': ['fido >= 4.2.1'],
         ':python_version<"3.5"': ['typing'],
+        'integration-tests': [
+            'bottle',
+            'ephemeral_port_reserve',
+            'pytest',
+        ]
     },
 )
